# docs/swagger.yaml
basePath: /
definitions:
...
  api.ErrorResponse:
    properties:
      details:
        example: '''to'' field is required'
        type: string
      error:
        example: validation failed
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@yourdomain.com
    name: API Support
  description: Unified notification delivery service supporting email, SMS, push notifications,
    and in-app messaging with bulk sending capabilities.
  title: Buzz Notification Service API
  version: 1.0.0
paths:
...
  /webhooks/generic:
    post:
      consumes:
      - application/json
      description: Receive and log generic webhook payloads
      parameters:
      - description: Webhook payload
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Generic webhook
      tags:
      - webhooks
  /webhooks/notifylk:
    post:
      consumes:
      - application/json
      description: Receive SMS delivery receipts from the NotifyLK provider
      parameters:
      - description: NotifyLK delivery receipt
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: NotifyLK webhook
      tags:
      - webhooks
  /webhooks/ses:
    post:
      consumes:
      - application/json
      description: Receive bounce, complaint, and delivery notifications from Amazon
        SES via SNS
      parameters:
      - description: SNS notification payload
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Amazon SES webhook
      tags:
      - webhooks
  /webhooks/twilio:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Receive SMS status callbacks from Twilio (form-encoded payload)
      parameters:
      - description: Twilio message SID
        in: formData
        name: MessageSid
        required: true
        type: string
      - description: Delivery status
        in: formData
        name: MessageStatus
        required: true
        type: string
      - description: Error code (if failed)
        in: formData
        name: ErrorCode
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Twilio webhook
      tags:
      - webhooks
securityDefinitions:
  Bearer:
    description: API key prefixed with "Bearer ", e.g. "Bearer YOUR_API_KEY"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"